requirements.txt
flask==3.0.0
python-dotenv==1.0.0
tweepy==4.14.0
openai==1.3.0
schedule==1.2.0
discord.py==2.3.2
python-telegram-bot==20.7